SFR-1513 Add Database Maintenance process
This adds a process that can perform "behind-the-scenes" maintenance of the database on a set schedule. Initially this performs a `VACUUM ANALYZE` operation on the main tables used in the database. This should provide a general performance improvement. It should be run on a weekly basis during the standard eReading maintenance window.

This can be extended to run further database operations.