Register background task apps for new relic
I *believe* this is the missing piece for covering our non web processes
for full instrumentation.  Basically, we just need to register an app
and pass that to the BackgroundTask context manager to get the newrelic
agent to fully instrument and forward logs from these tasks.

To test this, I performed the Docker based newrelic agent installation
on my machine, ran the ClassifyProcess locally, and observed logs and
metrics in the 'DRB ETL Pipeline (DEV)' app in newrelic APM.

Using this, looks like we don't need to individually decorate methods
being called in the task.  I only removed the decorators from the
ClassifyProcess for now so that we can first verify this in prod.

https://jira.nypl.org/browse/SFR-1603