Implement non search based automatic collection
As a first step, add the new models and implement the logic to fetch
editions for any collections not requiring text search. In these cases,
such as with the main example of a 'Most Recent' collection, we can just
hit the DB. This also includes a script to generate the 'Most Recent'
collection.

In a later diff, I'll implement handling of all the text search fields
that we are currenly ignoring.