Fix query for pagination and to dedup by work
Since limit might be null, we now want to properly implement pagination
in the actual SQL we generate, since we might now be requesting data
without an actual bound. To do that, pass the page parameters down so we
can build a limit / offset, and also refactor some code so that we don't
do some of the list slicing we do for static collections.

Also implement Mike's feedback to dedup editions on the work object,
using a RANK OVER PARTITION BY approach to fetch the earliest edition
per work object.