Remove `joinedload` from editions query
The joinedload seemed to add a bunch of extra results I think, maybe
because of all the outer joins it creates. I tried to use `.distinct` to
make a DISTNCT ON clause to dedup.

The query I *wanted* was this:
https://gist.github.com/sarangj/69263648c3fc76734fc5b56e8a592ee6 with
the DISTINCT ON at the top level.

However, because of how sqlalchemy does the joinedload stuff, it kept
stuffing my DISTINCT ON into a subquery, which defeated the purpose:
https://gist.github.com/sarangj/d0b283464e5f35efb12dd09436d1f8f1

So I just removed the joinedload, which appears to still allow the thing
to work? I assume sqlalchemy under the hood just lazily loads attributes
from the DB when they're accessed...