Pass collection into OPDS feed method
I noticed that we fetch the collection when calling this method even
though most of the time, we already have a collection available. This is
especially problematic in the `list` endpoint, where we end up
refetching each collection in the list. So instead, pass collection
objects into this OPDS feed method