Implement a publisher backlist cover fetcher
Add a new cover fetcher that looks for PDFs in DRBs publisher backlist
directory in s3 and tries to "infer" a cover from a PDF, by looking for
the first page in the PDF with a non-empty appearing image.

For determining "non-empty" images, we use the pillow package's entropy
value, since these images are scans, and the blank images that sometimes
are included in the PDF before the actual cover are not true blanks (ie,
there are minor variations in pixel color)

Also note, instead of using the S3 API to access the files, I opted to
use directly fetch the object urls using http requests, since the
objects are public, and so that this could potentially be adapted to
non-DRB hosted PDFs that are missing cover images.